crawler/requirements.txt
wheel>=0.37.1
bs4>=0.0.1
GitPython>=3.1.27
Jinja2>=3.1.2
PyYAML>=6.0
requests>=2.28.1
validators>=0.20.0
